+++ schema/site_stringdata.go
@@ -354,11 +354,17 @@ const SiteSchemaJSON = `{
       "group": "Search"
     },
     "search.limits": {
-      "description": "Limits that search applies for number of repositories searched.",
+      "description": "Limits that search applies for number of repositories searched and timeouts.",
       "type": "object",
       "group": "Search",
       "additionalProperties": false,
       "properties": {
+        "maxTimeoutSeconds": {
+          "description": "The maximum value for \"timeout:\" that search will respect. \"timeout:\" values larger than maxTimeoutSeconds are capped at maxTimeoutSeconds. Note: You need to ensure your load balancer / reverse proxy in front of Sourcegraph won't timeout the request for larger values. Note: Too many large rearch requests may harm Soucregraph for other users. Defaults to 1 minute.",
+          "type": "integer",
+          "default": "60",
+          "minimum": 1
+        },
         "maxRepos": {
           "description": "The maximum number of repositories to search across. The user is prompted to narrow their query if exceeded. Any value less than or equal to zero means unlimited.",
           "type": "integer",